Extract Reporter class
